refactor: changing folder structure
