+++ src/features/ui/components/Header/index.tsx
@@ -1,7 +1,7 @@
 import Link from 'next/link'
 import type { FC } from 'react'
 
-import { Routes } from '~/constants/routes'
+import { Routes } from '~/features/core/constants/routes'
 
 export const Header: FC = () => (
   <header>
