+++ src/pages/_app.tsx
@@ -1,4 +1,4 @@
-import '../theme/globals.css'
+import '~/features/ui/theme/globals.css'
 import type { AppProps } from 'next/app'
 
 function MyApp({ Component, pageProps }: AppProps) {
